DEV-005 refactor folder names to be singular
