+++ server/model/Mentor.js
@@ -16,15 +16,12 @@ const MentorSchema = new mongoose.Schema({
   },
 })
 
-
 MentorSchema.set('toJSON', {
   transform: (document, returnedObject) => {
     returnedObject.id = returnedObject._id.toString()
     delete returnedObject._id
     delete returnedObject.__v
-  }
+  },
 })
-
-
 
 module.exports = mongoose.model('Mentor', MentorSchema)
